backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Migrate==4.1.0
Pillow==10.1.0
pytesseract==0.3.10
reportlab==4.0.7
supabase==2.16.0
firebase-admin==6.4.0
python-dotenv==1.0.0
Werkzeug==3.0.1
opencv-python==4.10.0.84
numpy==1.26.4
icalendar==5.0.12
requests==2.31.0
pytz==2023.3 